### 2023-12-18 #6034 @Kevin-Mattheus-Moerman
@oleksulkowski I have just removed the `query-scope` label as it was determined this work may be in scope for JOSS. I will now proceed to find a handling editor.
