### 2024-01-27 #6034 @fboehm
@warrenlr @babinyurii @nterhoeven @ramadatta @oschwengers - would you be willing to review this submission for JOSS? As you know, our reviews are checklist-driven, meaning that a reviewer will be given a checklist of items to assess for the software and accompanying short manuscript.

Please feel free to ask me any questions. Thank you for considering this invitation.
