### 2024-01-28 #6034 @babinyurii
Hi, Frederick!
Yes, I can review this submission.

сб, 27 янв. 2024 г. в 18:58, Frederick Boehm ***@***.***>:

> @warrenlr <https://github.com/warrenlr> @babinyurii
> <https://github.com/babinyurii> @nterhoeven
> <https://github.com/nterhoeven> @ramadatta <https://github.com/ramadatta>
> @oschwengers <https://github.com/oschwengers> - would you be willing to
> review this submission for JOSS? As you know, our reviews are
> checklist-driven, meaning that a reviewer will be given a checklist of
> items to assess for the software and accompanying short manuscript.
>
> Please feel free to ask me any questions. Thank you for considering this
> invitation.
>
