### 2024-02-02 #6034 @fboehm
@warrenlr @nterhoeven @ramadatta @oschwengers - would you be willing to review this submission for JOSS? As you know, our reviews are checklist-driven, meaning that a reviewer will be given a checklist of items to assess for the software and accompanying short manuscript.

Please feel free to ask me any questions. Thank you for considering this invitation.
