### 2024-02-06 #6034 @fboehm
@KatyBrown @ns-rse @ammaraziz @will-rowe @dsurujon - would you be willing to review this submission for JOSS? Our reviews are checklist-driven, meaning that a reviewer will be given a checklist of items to assess for the software and accompanying short manuscript. Please feel free to ask me questions :) Thank you for considering this invitation!
