### 2024-02-07 #6332 @editorialbot
[REVIEW]: Syn-CpG-Spacer: A Panel web app for synonymous recoding of viral genomes with CpG dinucleotides
**Submitting author:** @oleksulkowski (<a href="http://orcid.org/0000-0002-0624-428X">Aleksander Sułkowski</a>)
**Repository:** https://github.com/oleksulkowski/Syn-CpG-Spacer
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
