### 2024-02-16 #6332 @babinyurii
Hi, @fboehm !
I've left unchecked `Functionality documentation`. I think docstring are needed in the `index.py` at least describing the classes and other objects.

Should I address other issues here which are not covered by these checkboxes? I've got some.
