### 2024-02-17 #6332 @fboehm
> Hi, @fboehm ! I've left unchecked `Functionality documentation`. I think docstring are needed in the `index.py` at least describing the classes and other objects.
>
> Should I address other issues here which are not covered by these checkboxes? I've got some.

Thank you @babinyurii - yes, please do offer suggestions that may improve the product. Please feel free to open new issues on the software repo's github page to describe the issues and offer suggestions.
