### 2024-02-20 #6332 @KatyBrown
> Hi, @fboehm ! I've left unchecked `Functionality documentation`. I think docstring are needed in the `index.py` at least describing the classes and other objects.

I agree with this and similarly comments would be useful in the JS wrapper in index.js
