### 2024-02-20 #6332 @KatyBrown
I'm not sure if this is outside of my scope as a reviewer, but I think it would be preferable to read the code in index.py - either directly or into a string - and then process this within index.js, if that's possible, rather than pasting it across. It seems risky to have to remember to paste the code and it brings the Python code which is actually running outside of the CI testing.
