### 2024-02-20 #6332 @oleksulkowski
Hi @KatyBrown, thank you for your feedback. The code in index.py is transferred to index.js when `panel convert index.py --to pyodide-worker --out docs/app --title Syn-CpG-Spacer --pwa` is run. If you want, I can emphasise that this command must always be run after changes are made to index.py in the "Development" section on the repository page.

@babinyurii could you please share the other issues? I will start working on your suggestions then.
