### 2024-02-20 #6332 @KatyBrown
> Hi @KatyBrown, thank you for your feedback. The code in index.py is transferred to index.js when `panel convert index.py --to pyodide-worker --out docs/app --title Syn-CpG-Spacer --pwa` is run. If you want, I can emphasise that this command must always be run after changes are made to index.py in the "Development" section on the repository page.
>
@oleksulkowski apologies, I missed that - in that case I think it’s fine as it is.
