### 2024-02-21 #6332 @KatyBrown
@fboehm OK, my review is complete now. Apart from extending the in-code documentation as @babinyurii discussed above, I think it looks good and meets all the requirements. It will be a useful tool and is well written and documented.
