+++ environment.yml
@@ -4,8 +4,8 @@ channels:
   - defaults
 dependencies:
   - python=3.11
-  - panel
-  - biopython
-  - bokeh
-  - pytest
+  - panel=1.3.8
+  - biopython=1.78
+  - bokeh=3.3.4
+  - pytest=7.4
   
