### 2024-02-26 64c2def
Fix issue #4 - detect .fas and .fa extensions
