+++ index.py
@@ -894,7 +894,7 @@ load_sample_btn = pnw.Button(
 )
 load_sample_btn.on_click(load_sample)
 
-file_input = pnw.FileInput(accept=".fasta,.aln")
+file_input = pnw.FileInput(accept=".fasta,.fa,.fas,.aln")
 
 
 # Define function to load FASTA file and display sequences
@@ -968,7 +968,7 @@ def load_fasta(event):
         successful_load_dummy.value = not successful_load_dummy.value
 
         file_input.param.unwatch(file_input_watcher)
-        new_file_input = pnw.FileInput(accept=".fasta,.aln")
+        new_file_input = pnw.FileInput(accept=".fasta,.fa,.fas,.aln")
         new_file_input_watcher = new_file_input.param.watch(load_fasta, "value")
         top[1] = new_file_input
         file_input = new_file_input
@@ -1146,7 +1146,7 @@ top_message = pn.pane.Markdown(
 
     Please load an in-frame sequence, of a length divisible by 3.
 
-    If loading a Fasta file with more than one sequence, ensure that \
+    If loading a FASTA file with more than one sequence, ensure that \
         sequence lengths are equal.
 
     In such case, the sequence at the top of the file will be used by the \
