### 2024-02-26 74ed2c6
Fix issue #5 - Clarify requirements for input file and warnings.
