+++ index.py
@@ -184,6 +184,37 @@ DIC_for_A_rich = {
 
 
 class Codon:
+    """
+    Represents a codon, which is a sequence of three nucleotides that together form a unit of genetic code in a DNA or RNA molecule.
+
+    Attributes:
+        position (int): The position of the codon within the Gene object, indicating the index of the first nucleotide of the codon.
+        sequence (str): The sequence of nucleotides in the codon.
+        next_codon (Codon, optional): Reference to the next codon object in the sequence, if applicable.
+        has_cg (bool): Indicates whether the codon contains a CpG dinucleotide.
+        protected (bool): Flag indicating whether the codon should not be mutated.
+        cg_split (bool): Indicates if a CpG dinucleotide is split over two subsequent codons.
+        mutable (bool, optional): Indicates if the codon can be mutated to contain a CpG dinucleotide.
+        potential_new_seq (str, optional): Holds a potential new sequence for the codon that might be considered for mutation, to form a CpG dinucleotide.
+        mutated (bool): Tracks whether the codon has been mutated.
+
+    Methods:
+        __init__(self, position, sequence, next_codon=None, has_cg=False, protected=False, cg_split=False, mutable=None, potential_new_seq=None, mutated=False):
+            Initializes a Codon object with the provided attributes.
+
+        get_cg_position(self, sequence):
+            Determines the position of a CpG dinucleotide within the codon or across adjacent codons, relative to the 'C'.
+
+        get_mutation_positions(self):
+            Identifies positions within the codon where mutations are possible based on the potential new sequence.
+
+        translate_codon(self):
+            Translates the codon sequence into its corresponding amino acid symbol according to the genetic code.
+
+        check_codon_viability(self):
+            Checks if the codon sequence is viable for translation based on its sequence.
+    """
+
     def __init__(
         self,
         position,
@@ -196,6 +227,21 @@ class Codon:
         potential_new_seq=None,
         mutated=False,
     ):
+        """
+        Initializes a Codon object with the provided attributes.
+
+        Parameters:
+            position (int): The position of the codon within the Gene object.
+            sequence (str): The nucleotide sequence of the codon.
+            next_codon (Codon, optional): The next codon in the sequence.
+            has_cg (bool): True if the codon contains a CpG dinucleotide, else False.
+            protected (bool): True if the codon is should not be mutated, else False.
+            cg_split (bool): True if a CpG dinucleotide is split between this codon and the next, else False.
+            mutable (bool, optional): True if the codon can potentially be mutated to contain a CpG, else False.
+            potential_new_seq (str, optional): A potential new sequence for mutation, forming a CpG dinucleotide.
+            mutated (bool): True if the codon has been mutated, else False.
+        """
+
         self.position = position  # Position of the codon, not the CG!
         self.sequence = sequence
         self.next_codon = next_codon
@@ -208,8 +254,17 @@ class Codon:
 
         self.check_codon_viability()
 
-    # Relative to the 'C'
     def get_cg_position(self, sequence):
+        """
+        Determines the position of a CpG dinucleotide within the codon or across adjacent codons, relative to the 'C'.
+
+        Parameters:
+            sequence (str): The nucleotide sequence of the codon to check for CpG presence.
+
+        Returns:
+            int or None: The position of the 'C' in the CpG dinucleotide, if present; otherwise, None.
+        """
+
         if sequence is not None:
             # Check for a CG pair within this codon
             for i in range(3):
@@ -228,6 +283,13 @@ class Codon:
         return
 
     def get_mutation_positions(self):
+        """
+        Identifies positions within the codon where mutations are possible based on the potential new sequence.
+
+        Returns:
+            list of int: A list of positions within the codon where mutations differ from the original sequence.
+        """
+
         diff_positions = []
 
[…]
