### 2024-02-27 #6332 @oleksulkowski
Hi, @babinyurii @KatyBrown, I have applied changes based on your suggestions - could you please check the issue pages and referenced commits?

I added the docstrings in commit oleksulkowski/Syn-CpG-Spacer@dc58e2a
