### 2024-03-05 #6332 @oleksulkowski
Hi @fboehm, I've completed the tasks. The new release version is [v1.1.0](https://github.com/oleksulkowski/Syn-CpG-Spacer/releases/tag/v1.1.0). The Zenodo DOI is https://doi.org/10.5281/zenodo.10781374.
