### 2024-03-14 #6332 @fboehm
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1146/annurev-virology-091919-104213 is OK
> - 10.1186/1756-0500-5-50 is OK
> - 10.18129/B9.bioc.SynMut is OK
> - 10.1371/journal.ppat.1000079 is OK
> - 10.1042/BST20200695 is OK
> - 10.1128/spectrum.02529-23 is OK
> - 10.1371/journal.ppat.1011357 is OK
> - 10.7554/eLife.12735 is OK
> - 10.1038/s41564-022-01223-8 is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: Attenuation of Human Respiratory Viruses by Synony...
>
> INVALID DOIs
>
> - None
> ```
@oleksulkowski - can you check on this? Is there a reason that it doesn't have a doi?
