### 2024-03-14 ec16913
fix bibliography
