+++ paper.bib
@@ -38,7 +38,7 @@ Publisher: {BioMed} Central},
 }
 
 @software{gu_synmut_2023,
-	title = {{SynMut}: {SynMut}: Designing Synonymously Mutated Sequences with Different Genomic Signatures},
+	title = {{SynMut}: Designing Synonymously Mutated Sequences with Different Genomic Signatures},
 	rights = {{GPL}-2},
 	url = {https://bioconductor.org/packages/SynMut/},
 	shorttitle = {{SynMut}},
@@ -76,6 +76,7 @@ Publisher: {BioMed} Central},
 	volume = {10},
 	issn = {1664-3224},
 	url = {https://www.frontiersin.org/articles/10.3389/fimmu.2019.01250},
+	doi = {10.3389/fimmu.2019.01250},
 	abstract = {Using computer algorithms and commercial {DNA} synthesis, one or more {ORFs} of a microbial pathogen such as a virus can be recoded and deoptimized by several strategies that may involve the introduction of up to thousands of nucleotide (nt) changes without affecting amino acid (aa) coding. The synonymous recoding strategies that have been applied to {RNA} viruses include: deoptimization of codon or codon-pair usage, which may reduce protein expression among other effects; increased content of immunomodulatory {CpG} and {UpA} {RNA}, which increase immune responses and thereby restrict viral replication; and substitution of serine and leucine codons with synonymous codons for which single-nt substitutions can yield nonsense codons, thus limiting evolutionary potential. This can reduce pathogen fitness and create potential live-attenuated vaccines that may have improved properties. The combined approach of genome recoding, synthetic biology, and reverse genetics offers several advantages for the generation of attenuated {RNA} viruses. First, synonymous recoding involves many mutations, which should reduce the rate and magnitude of de-attenuation. Second, increasing the amount of recoding can provide increased attenuation. Third, because there are no changes at the aa level, all of the relevant epitopes should be expressed. Fourth, attenuation frequently does not compromise immunogenicity, suggesting that the recoded viruses have increased immunogenicity per infectious particle. Synonymous deoptimization approaches have been applied to two important human viral pathogens, namely respiratory syncytial virus ({RSV}) and influenza A virus ({IAV}). This manuscript will briefly review the use of these different methods of synonymous recoding to generate attenuated {RSV} and {IAV} strains. It also will review the characterization of these vaccine candidates in vitro and in animal models, and describe several surprising findings with respect to phenotypic and genetic instability of some of these candidates.},
 	journaltitle = {Frontiers in Immunology},
 	author = {Le Nouën, Cyril and Collins, Peter L. and Buchholz, Ursula J.},
