### 2024-03-23 #6332 @fboehm
I just read the generated pdf, and I really like it. My only real issue is that I want you to cite more of the software packages that you mention:

1. in line 19, please add a reference for the "Panel" library.
2. in line 23, please add a reference for the "Bokeh" library.
3. in line 25, please add a reference for the "Biopython" suite of modules.
4. Is there a reference (or webpage) that can be referenced when describing "PWA"? If so, please cite it.
5. Pyodide, line 27, reference needed.

I hope that this is reasonable - please let me know if issues or concerns arise, @oleksulkowski
