### 2024-03-25 db56c17
Add more software references
