+++ paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 
 Vertebrate genomes contain lower than expected frequencies of the CpG dinucleotide. Consequently, many vertebrate viruses have evolved to mimic this composition, possibly in order to evade host antiviral defences [@greenbaum_patterns_2008]. For example, the antiviral protein ZAP binds CpGs in viral single stranded RNA with specific spacing requirements [@goncalves-carneiro_rational_2022], though CpGs are also likely depleted in viral genomes due to other selective pressures [@forni_dinucleotide_2023]. Increasing CpG abundance by synonymous recoding could facilitate attenuation of viruses without compromising their epitope antigenicity by changing non-CpG codons to alternatives containing CpG without changing the overall amino acid sequence [@le_nouen_attenuation_2019; @goncalves-carneiro_rational_2022; @sharp_cpg_2023]. There are three ways CpGs can be synonymously introduced in codons: at positions 1-2 for arginine (e.g. AGA → CGA), 2-3 for several amino acids (e.g. ACA → ACG), or in a 3-1 split configuration, if a subsequent codon begins with a G (e.g. ATA-GCA → ATC-GCA).
 
-Syn-CpG-Spacer is a Python progressive web app (PWA) made with the Panel library that allows for consistent recoding of viral sequences and applying biologically relevant constraints. These include setting a minimum gap between CpG’s, optimizing for an average CpG gap, protecting cis-acting regulatory signals from modification, and modulating the A-content in the overall sequence. The app features a sequence viewer made with the Bokeh library that highlights CpG dinucleotides, allowing for efficient analysis of the resulting distribution of CpGs. This is complemented by a statistical data table. Utilizing Biopython modules, the user can load their sequence as a FASTA file and download the outputs as an alignment in the same format. As a PWA running on Pyodide, the code is only executed in the user’s browser and they can install the app onto their machine for offline use.
+Syn-CpG-Spacer is a Python progressive web app (PWA) [@mdn_web_docs_progressive_2023] made with the Panel library [@panel_development_team_holovizpanel_2024] that allows for consistent recoding of viral sequences and applying biologically relevant constraints. These include setting a minimum gap between CpG’s, optimizing for an average CpG gap, protecting cis-acting regulatory signals from modification, and modulating the A-content in the overall sequence. The app features a sequence viewer made with the Bokeh library [@bokeh_development_team_bokeh_2014] that highlights CpG dinucleotides, allowing for efficient analysis of the resulting distribution of CpGs. This is complemented by a statistical data table. Utilizing Biopython [@cock_biopython_2009] modules, the user can load their sequence as a FASTA file and download the outputs as an alignment in the same format. As a PWA running on Pyodide [@the_pyodide_development_team_pyodidepyodide_2023], the code is only executed in the user’s browser and they can install the app onto their machine for offline use.
 
 # Statement of need
 
