+++ paper.bib
@@ -171,3 +171,65 @@ Publisher: {BioMed} Central},
 Publisher: Nature Publishing Group},
 	keywords = {Live attenuated vaccines, Restriction factors},
 }
+
+@software{bokeh_development_team_bokeh_2014,
+	title = {Bokeh: Interactive Data Visualization in the browser, from Python},
+	url = {https://github.com/bokeh/bokeh},
+	shorttitle = {bokeh/bokeh},
+	abstract = {Interactive Data Visualization in the browser, from  Python - bokeh/bokeh},
+	author = {Bokeh Development Team},
+	urldate = {2024-03-25},
+	date = {2014},
+}
+
+@article{cock_biopython_2009,
+	title = {Biopython: freely available Python tools for computational molecular biology and bioinformatics},
+	volume = {25},
+	issn = {1367-4803},
+	url = {https://doi.org/10.1093/bioinformatics/btp163},
+	doi = {10.1093/bioinformatics/btp163},
+	shorttitle = {Biopython},
+	abstract = {Summary: The Biopython project is a mature open source international collaboration of volunteer developers, providing Python libraries for a wide range of bioinformatics problems. Biopython includes modules for reading and writing different sequence file formats and multiple sequence alignments, dealing with 3D macro molecular structures, interacting with common tools such as {BLAST}, {ClustalW} and {EMBOSS}, accessing key online databases, as well as providing numerical methods for statistical learning.Availability: Biopython is freely available, with documentation and source code at www.biopython.org under the Biopython license.Contact: All queries should be directed to the Biopython mailing lists, see www.biopython.org/wiki/\_Mailing\_listspeter.cock@scri.ac.uk.},
+	pages = {1422--1423},
+	number = {11},
+	journaltitle = {Bioinformatics},
+	shortjournal = {Bioinformatics},
+	author = {Cock, Peter J. A. and Antao, Tiago and Chang, Jeffrey T. and Chapman, Brad A. and Cox, Cymon J. and Dalke, Andrew and Friedberg, Iddo and Hamelryck, Thomas and Kauff, Frank and Wilczynski, Bartek and de Hoon, Michiel J. L.},
+	urldate = {2024-03-25},
+	date = {2009-06-01},
+}
+
+@online{mdn_web_docs_progressive_2023,
+	title = {Progressive web apps},
+	url = {https://developer.mozilla.org/en-US/docs/Web/Progressive_web_apps},
+	abstract = {A progressive web app ({PWA}) is an app that's built using web platform technologies, but that provides a user experience like that of a platform-specific app.},
+	author = {{MDN} Web Docs},
+	urldate = {2024-03-25},
+	date = {2023-10-25},
+	langid = {american},
+}
+
+@software{the_pyodide_development_team_pyodidepyodide_2023,
+	title = {pyodide/pyodide: 0.24.1},
+	url = {https://zenodo.org/records/8378122},
+	shorttitle = {pyodide/pyodide},
+	abstract = {Pyodide is a Python distribution for the browser and Node.js based on {WebAssembly}},
+	version = {0.24.1},
+	publisher = {Zenodo},
+	author = {The Pyodide development team},
+	urldate = {2024-03-25},
+	date = {2023-09-25},
+	doi = {10.5281/zenodo.8378122},
+}
+
+@software{panel_development_team_holovizpanel_2024,
+	title = {holoviz/panel: Version 1.3.8},
+	url = {https://zenodo.org/records/10563859},
+	shorttitle = {holoviz/panel},
+	version = {v1.3.8},
+	publisher = {Zenodo},
+	author = {Panel Development Team},
+	urldate = {2024-03-25},
+	date = {2024-01-24},
+	doi = {10.5281/zenodo.10563859},
+}
