### 2024-03-25 b9b2afc
fix references authors
