+++ paper.bib
@@ -177,7 +177,7 @@ Publisher: Nature Publishing Group},
 	url = {https://github.com/bokeh/bokeh},
 	shorttitle = {bokeh/bokeh},
 	abstract = {Interactive Data Visualization in the browser, from  Python - bokeh/bokeh},
-	author = {Bokeh Development Team},
+	author = {{Bokeh Development Team}},
 	urldate = {2024-03-25},
 	date = {2014},
 }
@@ -203,32 +203,32 @@ Publisher: Nature Publishing Group},
 	title = {Progressive web apps},
 	url = {https://developer.mozilla.org/en-US/docs/Web/Progressive_web_apps},
 	abstract = {A progressive web app ({PWA}) is an app that's built using web platform technologies, but that provides a user experience like that of a platform-specific app.},
-	author = {{MDN} Web Docs},
+	author = {{MDN Web Docs}},
 	urldate = {2024-03-25},
 	date = {2023-10-25},
 	langid = {american},
 }
 
 @software{the_pyodide_development_team_pyodidepyodide_2023,
-	title = {pyodide/pyodide: 0.24.1},
+	title = {pyodide/pyodide: a Python distribution for the browser and Node.js based on WebAssembly},
 	url = {https://zenodo.org/records/8378122},
 	shorttitle = {pyodide/pyodide},
 	abstract = {Pyodide is a Python distribution for the browser and Node.js based on {WebAssembly}},
 	version = {0.24.1},
 	publisher = {Zenodo},
-	author = {The Pyodide development team},
+	author = {{The Pyodide development team}},
 	urldate = {2024-03-25},
 	date = {2023-09-25},
 	doi = {10.5281/zenodo.8378122},
 }
 
 @software{panel_development_team_holovizpanel_2024,
-	title = {holoviz/panel: Version 1.3.8},
+	title = {holoviz/panel: The powerful data exploration & web app framework for Python},
 	url = {https://zenodo.org/records/10563859},
 	shorttitle = {holoviz/panel},
 	version = {v1.3.8},
 	publisher = {Zenodo},
-	author = {Panel Development Team},
+	author = {{Panel Development Team}},
 	urldate = {2024-03-25},
 	date = {2024-01-24},
 	doi = {10.5281/zenodo.10563859},
