### 2024-03-25 #6332 @oleksulkowski
Thank you for pointing that out @fboehm. I've now added the references with DOIs, if available.
For the two references with missing DOIs, one is a website and the other is a GitHub repository.
