### 2024-03-28 #6332 @fboehm
@oleksulkowski - are you using the proper citation for Bokeh? I found this: https://docs.bokeh.org/en/1.0.4/docs/citation.html#:~:text=To%20cite%20Bokeh%20in%20publications,.bokeh.pydata.org.
