### 2024-03-28 #6332 @fboehm
@oleksulkowski - I can't quickly find another way to cite Progressive Web Apps, so, unless you know of one, we can leave it as is.
