+++ paper.bib
@@ -172,14 +172,15 @@ Publisher: Nature Publishing Group},
 	keywords = {Live attenuated vaccines, Restriction factors},
 }
 
-@software{bokeh_development_team_bokeh_2014,
-	title = {Bokeh: Interactive Data Visualization in the browser, from Python},
-	url = {https://github.com/bokeh/bokeh},
-	shorttitle = {bokeh/bokeh},
+@software{bokeh_development_team_bokeh_2024,
+	title = {Bokeh: Python library for interactive visualization},
+	url = {https://docs.bokeh.org/en/3.3.4/},
+	shorttitle = {Bokeh},
 	abstract = {Interactive Data Visualization in the browser, from  Python - bokeh/bokeh},
+	version = {3.3.4},
 	author = {{Bokeh Development Team}},
-	urldate = {2024-03-25},
-	date = {2014},
+	urldate = {2024-03-28},
+	date = {2024-01},
 }
 
 @article{cock_biopython_2009,
