### 2024-03-28 #6332 @oleksulkowski
I changed the Bokeh reference to include the version we used now. Regarding the PWA reference, the date accessed is there in the paper.bib file (`urldate`) but I think that the JOSS compiler doesn't show it on the pdf.
