### 2024-03-31 #6332 @Kevin-Mattheus-Moerman
@oleksulkowski as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Everything seems in order, I only have the below point that requires your attention:

- [ ] Most of your text in the paper features British English, except for `optimizing` and `utilizing`, you may consider editing these to be consistent.
